refactor: move App para um arquivo próprio
