+++ lib/main.dart
@@ -1,32 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'app.dart';
 import 'shared/dependencies.dart';
-import 'shared/palette.dart';
-import 'pages/home_page.dart';
 
 void main() {
   registerDependencies();
 
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return ProviderScope(
-      child: MaterialApp(
-        debugShowCheckedModeBanner: false,
-        title: 'Arara',
-        theme: ThemeData(
-          primarySwatch: Palette.paletaRoxa,
-          visualDensity: VisualDensity.adaptivePlatformDensity,
-        ),
-        home: const MyHomePage(),
-      ),
-    );
-  }
+  runApp(const App());
 }
